% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wcvp_checklist_distribution}
\alias{wcvp_checklist_distribution}
\title{The World Checklist of Vascular Plants: distributions}
\format{
A data frame with 1,990,342 rows and 11 variables:
\describe{
\item{plant_locality_id}{Sequential number.}
\item{plant_name_id}{WCVP Identifier, corresponding to the filed of the same name in checklist_names.txt.}
\item{continent_code_l1}{Numerical botanical continent code (TDWG level 1).}
\item{continent}{Botanical continent (TDWG Level 1). See \url{https://www.tdwg.org/standards/wgsrpd/} for details.}
\item{region_code_l2}{Numerical botanical region code (TDWG level 2).}
\item{region}{Botanical region (TDWG Level 2). See \url{https://www.tdwg.org/standards/wgsrpd/} for details.}
\item{area_code_l3}{Numerical botanical area code (TDWG level 3).}
\item{area}{Botanical area (TDWG Level 3). See \url{https://www.tdwg.org/standards/wgsrpd/} for details.}
\item{introduced}{Introduced status of the taxon: 0 if native; 1 if introduced.}
\item{extinct}{Local extinction status of the taxon: 0 if extant; 1 if extinct in region.}
\item{location_doubtful}{Trustfulness of status: 1 if doubtfully present; else 0.}
}
}
\source{
\url{http://sftp.kew.org/pub/data-repositories/WCVP/wcvp.zip}
}
\usage{
wcvp_checklist_distribution
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
A dataset containing the distribution data from the WCVP, mapped to the
Biodiversity Information Standards (TDWG) World Geographical Scheme for
Recording Plant Distributions (WGSRPD)
}
\examples{
\donttest{
data(wcvp_checklist_distribution)
head(wcvp_checklist_distribution)
}
}
\keyword{datasets}
